backend/rag/requirements.txt
langchain==0.3.20
langchain-openai==0.3.9
langchain-community==0.3.19
langchain-core==0.3.45
langchain-chroma==0.2.2
pydantic==2.10.6
python-dotenv==1.0.1
openai==1.66.3
sentence-transformers==3.4.1
chromadb==0.6.3
tiktoken==0.9.0
scikit-learn==1.6.1
numpy==1.26.4
pypdf==5.3.1
rank-bm25==0.2.2
fastapi==0.115.11
uvicorn==0.34.0
